Task 1 - Base Structure Created
This is the html structure that includes the header, the main body section, and the footer. The header is labeled as "Ticket Information" and the footer contains copyright information as a filler. Within the main section we have created a div container with the id equal to "ticketContainer" that will be referenced to later in the javascript file that is also linked to the html file.