Task 2 - Support Tickets Dynamic Addition
Create a function that plugs in the customer name, a description of the issue, and the priority level into the html display that is linked to each support ticket. Each ticket has a resolve button that removes the specific ticket indicating that the support ticket is no longer open. Each support ticket is linked to the the ticket container.