Task 3 - Highlighting High Priority Tickets
I edited the code so that priority level is set as an attribute of each support ticket that way when I use querySelectorAll, I am able to reference only the high priority level tickets and store them in an array. For each ticket in the array, I changed the text color to a maroon and the background color to a light pink.