Task 4 - Support Ticket Resolution with Event Bubbling
Implement removeChild to the "resolve" button that removes the parent ticket as well as stopPropagation() to prevent clicking the button to also print "ticket clicked" to the console. Make it so that the addEventListener recognizes only when a ticket is clicked and logs the "ticket clicked" message to the console.