Task 5 - Inline Editing for Support Tickets
Creating a function that allows one to edit the text within the browser for the customer name, the issue description, and the priority level and save it. This is done by establishing an edit button or double clicking the support ticket to open the editing mode and a saving button that will replace the text and exit out of the edit mode.